% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{char_2_dot}
\alias{char_2_dot}
\title{char_2_dot}
\usage{
char_2_dot(char)
}
\arguments{
\item{char}{character}
}
\description{
char_2_dot
}
\examples{
db <- dplyr::lahman_sqlite()
db_names <- get_names(db)["AllstarFull"]

char_2_dot(db_names)
}

