% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{connectives}
\alias{conjunct}
\alias{connectives}
\alias{disjunct}
\title{Conjunction and disjunction}
\usage{
conjunct(..., where = parent.frame())

disjunct(..., where = parent.frame())
}
\arguments{
\item{...}{Predicate functions whose conjunction or disjunction to return.}

\item{where}{An environment in which to resolve any character or
symbol arguments.}
}
\value{
A function of a single argument, representing the conjunction or
disjunction of the predicates' values on its argument.
}
\description{
\code{conjunct} and \code{disjunct} construct and return functions which
represent the conjunction or disjunction of predicates.
}
\details{
The functions passed must all take a single argument.

If some of the \code{...} arguments are not function objects, they are
resolved to function objects in the following way: character arguments
are looked up in the provided \code{where} environment; symbol arguments
are coerced to character and then looked up in the \code{where} environment;
other types of arguments, or symbol/character arguments which fail to
resolve to a function object, raise an error.
}
\examples{
f <- conjunct(is.positive, is.odd)
g <- conjunct(is.positive, is.zero)
h <- disjunct(is.negative, is.odd)

f(3) == TRUE
f(2) == FALSE

#false for all x - no positive numbers are 0
g(0) == FALSE
g(10) == FALSE

h(10) == FALSE
h(-10) == TRUE
h(3) == TRUE

}
\seealso{
\code{compose}, which returns the composition rather than the conjunction or
disjunction.
}

