% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{export_svg}
\alias{export_svg}
\title{export_svg}
\usage{
export_svg(rendered, fpath)
}
\arguments{
\item{rendered}{output from \code{\link[schemeR]{render_sqlite}} or \code{\link[schemeR]{render_df}}}

\item{fpath}{file.path}
}
\description{
export_svg
}
\details{
https://github.com/rich-iannone/DiagrammeR/issues/133
}

