% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{for.each}
\alias{for.each}
\title{Zipped function application}
\usage{
for.each(f, ...)
}
\arguments{
\item{f}{A function to apply.}

\item{...}{Vectors, lists or pairlists.}
}
\value{
The value of f on the last set of parameters supplied to it, or
NULL if no or only zero-length sequences were supplied as arguments.
}
\description{
\code{for.each} takes a function and a set of sequences (vectors, lists or
pairlists), and applies the function successively to each set of elements
that are in the sequences at the same positional index.
}
\details{
This function combines \code{link{zip}} and \code{link{do.call}}, to apply
a provided function successively to the elements of several sequences. The
arity of the function must be compatible with the number of sequences
passed as arguments; if not, an error will result.

As in \code{\link{zip}}, if not all sequences are the same length, shorter
sequences will be recycled to match the lengths of longer ones.
}
\examples{
f <- function(...) print(as.numeric(list(...)))
invisible(for.each(f, 1:5, 6:10, 11:15))

}
\seealso{
\code{link{zip}} for the operation of zipping the parameter sequences.
}

