% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{gensym}
\alias{gensym}
\title{Build a temporary symbol}
\usage{
gensym(str = "G", envir = parent.frame(), len = 10)
}
\arguments{
\item{str}{A string to prepend to the generated symbol name.}

\item{envir}{The environment the symbol should be unique in.}

\item{len}{How long (in characters) the symbol should be.}
}
\value{
The generated symbol.
}
\description{
\code{gensym} is an analogue of Common Lisp's gensym. It builds symbols
which are guaranteed not to shadow or conflict with bindings already defined
in an environment, for use in writing macros.
}
\details{
\code{gensym} provides some additional control over the form of the symbol
it generates: the user can specify how long the symbol should be (though
asking for length-1 unique symbols is unlikely to be useful), and ask for a
particular string to be prepended for ease of processing or as even more
insurance against name conflicts (perhaps relative to other environments).
}

