% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{get_names}
\alias{get_names}
\title{get_names}
\usage{
get_names(sqlite)
}
\arguments{
\item{sqlite}{SQLite database}
}
\description{
get_names
}
\examples{
\dontrun{
get_names(dplyr::tbl(dt, "TeamsHalf"))
}
}

