% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{infix}
\alias{infix}
\title{Convert prefix-formatted R code to infix}
\usage{
infix(expr)
}
\arguments{
\item{expr}{A prefix-formatted expression (via the \code{.} function).}
}
\value{
The prefix-formatted expression converted to infix.
}
\description{
This function converts R code written in a Lisp-like prefix format, via the
\code{\link{.}} function, to the usual infix R syntax. A call to infix()
does not evaluate the generated expression, partly for efficiency reasons,
though it does partial evaluation of the passed expression: each call to
\code{.} is evaluated to the corresponding infix call.
}
\section{See Also}{

The \code{\link{.}} function, which is the building block of prefix-
formatted R code; \code{\link{prefix}}, the inverse of \code{infix};
\code{\link{infixFiles}} and \code{\link{prefixFiles}} for converting
code found in files.
}

