% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{infixFiles}
\alias{infixFiles}
\title{Convert one or more R code files from prefix to infix}
\usage{
infixFiles(files, overwrite = FALSE)
}
\arguments{
\item{files}{A character vector of file paths to convert.}

\item{overwrite}{Should the generated code overwrite the input file?}
}
\value{
Invisible NULL.
}
\description{
This function converts files from prefix to infix. It can be told to
overwrite the passed files, or to generate new files. In the
latter case, the input files are assumed to have extension ".Rl", and the
corresponding output files will have extension ".R".
}
\section{See Also}{

\code{\link{prefixFiles}} for converting files in the other direction;
\code{\link{infix}} and \code{\link{prefix}} for converting expressions
from R-style infix to Lisp-style prefix and vice versa; the \code{\link{.}}
function, which is the building block of prefix-formatted R code.
}

