% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{macro}
\alias{defmacro}
\alias{macro}
\title{Lisp macros for R}
\usage{
defmacro(...)

macro(...)
}
\arguments{
\item{...}{The infix form of prefix arguments.}
}
\value{
Macro returns the created macro, which is an R function. Defmacro,
as in Common Lisp, returns the symbol it's bound the new macro to.
}
\description{
foo
}
\details{
bar
This is the non-hygienic version of macros, based on defmacro() as in
Common Lisp, rather than on syntax objects and syntax-case.
}

