% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{prefixFiles}
\alias{prefixFiles}
\title{Convert one or more R code files from infix to prefix}
\usage{
prefixFiles(files, overwrite = FALSE)
}
\arguments{
\item{files}{A character vector of file paths to convert.}

\item{overwrite}{Should the generated code overwrite the input file?}
}
\value{
Invisible NULL.
}
\description{
This function converts files from infix to prefix. It can be told to
overwrite the passed files, or to generate new files. In the
latter case, the input files are assumed to have extension ".R", and the
corresponding output files will have extension ".Rl".
}
\section{See Also}{

\code{\link{infixFiles}} for converting files in the other direction;
\code{\link{infix}} and \code{\link{prefix}} for converting expressions
from R-style infix to Lisp-style prefix and vice versa; the \code{\link{.}}
function, which is the building block of prefix-formatted R code.
}

