% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix.R
\name{quasiquote}
\alias{quasiquote}
\title{Partial quoting in expressions}
\usage{
quasiquote(expr, where = parent.frame())
}
\arguments{
\item{expr}{The expression to be partially quoted.}

\item{where}{The environment in which any evaluation should occur.}
}
\value{
The quoted expr, with partial evaluation and substitution done.
}
\description{
A reimplementation of the backquote macro from Lisp/Scheme, with more
functionality than the base \code{bquote} function. quasiquote quotes
its argument, except for two kinds of user-requested evaluation and
interpolation. Adapted from base R's \code{bquote}.
}
\details{
Unlike the bquote() found in base R, this version handles both of the types
of unquoting Lisp provides:
\itemize{
\item{terms wrapped in \code{.c()} are evaluated in the environment given
by \code{where}. This feature replicates Lisp's \code{unquote}, or the
comma notation, and behaves identically to \code{bquote}'s \code{.()}.}
\item{terms wrapped in \code{.s()} are evaluated in the environment given
by \code{where}, are assumed to evaluate to a list or pairlist, and the
elements of that list or pairlist are spliced into the expression where
the \code{.s()} occurred. This feature replicates Lisp's
\code{unquote-splicing}, or the comma-at notation.}
}
}
\section{Warning}{

Splicing unquote via \code{.s()} works on a purely lexical basis, and unlike
in Lisp, there is no guarantee the resulting object will make any sense.
}

