% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_df}
\alias{render_df}
\title{render_df}
\usage{
render_df(df)
}
\arguments{
\item{df}{}
}
\description{
render_df
}
\examples{
df <- data.frame(
         "name" = 1:4,
         "people" = c("bob", "jim", "beatrice", "ann"))
         render_df(df)

}

