% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{render_sqlite}
\alias{render_sqlite}
\title{render_sqlite}
\usage{
render_sqlite(db)
}
\arguments{
\item{db}{SQLite database}
}
\description{
render_sqlite
}
\examples{

render_sqlite(dplyr::lahman_sqlite())
}

