% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheme.R
\name{schemeR}
\alias{schemeR}
\title{A custome ggplot2 theme for styles of figures that I like.}
\usage{
schemeR()
}
\arguments{
\item{graph_type}{The type of graph for the theme}

\item{angle}{Angle for X-axis labels}

\item{base_size}{Pt. size for text}

\item{base_family}{Font type, I think.}

\item{base_line_size}{Size of lines, default is base_size/22}

\item{base_rect_size}{Size of rect, default is base_size/22}
}
\value{
ggplot-theme
}
\description{
A custome ggplot2 theme for styles of figures that I like.
}
