% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional.R
\name{zip}
\alias{zip}
\title{Transposition of sequences}
\usage{
zip(...)
}
\arguments{
\item{...}{Sequences to zip.}
}
\value{
The zipped sequences.
}
\description{
\code{zip} takes lists, vectors or pairlists and creates new ones that
represent the transposition of the originals.
}
\details{
"Transposition" here is meant in the following sense: given n sequences
of length m, return m sequences of length n. This sounds confusing but
is quite intuitive in practice - see the examples.

For those familiar with Python, this is just like the built-in Python
\code{zip} function.
}
\examples{
f <- 1:5
g <- 5:10
h <- 11:15

zip(f, g, h)

}

