% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemeR.R
\name{df_2_dot}
\alias{df_2_dot}
\title{df_2_dot}
\usage{
df_2_dot(df_name, df)
}
\arguments{
\item{df}{}
}
\description{
df_2_dot
}
\examples{
df <- data.frame(
         "name" = 1:4,
         "people" = c("bob", "jim", "beatrice", "ann"))
df_2_dot("df", df)
}

