% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defmacro.R
\name{gensym}
\alias{gensym}
\title{Build a temporary symbol}
\usage{
gensym(str = "G", len = 10)
}
\arguments{
\item{str}{A string to prepend to the generated symbol name.}

\item{len}{How long (in characters) the symbol should be.}
}
\value{
The generated symbol.
}
\description{
\code{gensym} is an analogue of Common Lisp's gensym, for use in writing
macros. It builds randomly generated symbols which can be made to avoid
shadowing bindings already defined in an environment.
}
\details{
\code{gensym} provides some additional control over the form of the symbol
it generates: the user can specify how long the symbol should be (though
asking for length-1 unique symbols is unlikely to be useful), and ask for a
particular string to be prepended for ease of processing or as even more
insurance against name conflicts.
}
\seealso{
The main use of this function is providing temporary symbols that macro
definitions can use to avoid capturing variables from the calling
environment; see \code{\link{macro}} and \code{link{defmacro}} for the macro
facility in question.
}

