% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{prefix}
\alias{prefix}
\title{Convert infix-formatted R code to prefix}
\usage{
prefix(expr)
}
\arguments{
\item{expr}{An infix-formatted expression.}
}
\value{
The infix-formatted expression converted to prefix.
}
\description{
This function converts R code written in the usual infix R syntax to
a Lisp-like prefix syntax expressed via the \code{\link{.}} function. A
call to prefix() does not evaluate the passed expression or the generated
one, partly for efficiency reasons.
}
\section{See Also}{

The \code{\link{.}} function, which is the building block of prefix-
formatted R code; \code{\link{infix}}, the inverse of \code{prefix};
\code{\link{infixFiles}} and \code{\link{prefixFiles}} for converting
code found in files.
}

