% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoting.R
\name{quasiquote}
\alias{quasiquote}
\title{Partial quoting in expressions}
\usage{
quasiquote(expr, where = parent.frame())
}
\arguments{
\item{expr}{The expression to be partially quoted.}

\item{where}{The environment in which any evaluation should occur.}
}
\value{
The quoted expr, with partial evaluation and substitution done.
}
\description{
A reimplementation of the backquote macro from Lisp/Scheme, with more
functionality than the base \code{bquote} function. quasiquote quotes
its argument, except for two kinds of user-requested evaluation and
interpolation. Adapted from base R's \code{bquote}.
}
\details{
Unlike the bquote() found in base R, this version handles both of the types
of unquoting Lisp provides:
\itemize{
\item{terms wrapped in \code{.c()} are evaluated in the environment given
by \code{where}. This feature replicates Lisp's \code{unquote}, or the
comma notation, and behaves identically to \code{bquote}'s \code{.()}.}
\item{terms wrapped in \code{.s()} are evaluated in the environment given
by \code{where}, are assumed to evaluate to a list, pairlist or vector,
and its elements are spliced into the expression where the \code{.s()}
occurred. This feature replicates Lisp's \code{unquote-splicing}, or the
comma-at notation.}
}
}
\section{Warning}{

Splicing unquote via \code{.s()} works on a purely lexical basis, and unlike
in Lisp, there is no guarantee the resulting object will make any sense.
}
\examples{
f <- list(1,2,3)
eval(quasiquote(list(.c(f), 4, 5, 6)))

f <- list(5,6,7)
eval(quasiquote(list(1, 2, 3, 4, .s(f), 8, 9, 10)))

}

